requirements.txt
absl-py==0.9.0
appdirs==1.4.4
asgiref==3.2.10
astor==0.8.1
astunparse==1.6.3
beautifulsoup4==4.9.1
bleach==1.5.0
bs4==0.0.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
distlib==0.3.0
Django==3.0.8
enum34==1.1.10
filelock==3.0.12
gast==0.3.3
google-auth==1.16.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
gspread==3.6.0
h5py==2.10.0
html5lib==0.9999999
httplib2==0.18.1
idna==2.9
importlib-metadata==1.6.0
importlib-resources==1.5.0
Keras-Preprocessing==1.1.2
Markdown==3.2.2
numpy==1.16.4
oauth2client==4.1.3
oauthlib==3.1.0
opt-einsum==3.2.1
Pillow==7.2.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==1.4.0
tensorflow-estimator==2.2.0
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
urllib3==1.25.9
virtualenv==20.0.21
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
